Rename "normal tokens" to "fungible tokens"
