+++ chainstate/test-suite/src/tests/fungible_tokens.rs
